# lib/api-zod/src/generated/types/bloomAnnotation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom-aware Personalized Lesson Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface BloomAnnotation {
  primary_bloom_level: string;
  secondary_levels?: string[];
  confidence: number;
  reasoning: string;
  keywords_matched?: string[];
}